requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
yfinance>=0.2.54
PyYAML>=6.0
websockets>=12.0
aiohttp>=3.10.0
